Refactor all workflow utils into common file
